+++ src/components/app/app.tsx
@@ -1,14 +1,14 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Main from '../pages/main';
-import Login from '../pages/login';
-import Offer from '../pages/offer';
-import NotFound from '../pages/not-found';
-import Favorites from '../pages/favorites';
-import PrivateRoute from './private-route';
-import { useAppSelector } from '../hooks';
-import LoadingScreen from '../pages/loading-screen';
-import Layout from './layout';
-import { AuthStatus } from '../utils/const';
+import Main from '../../pages/main';
+import Login from '../../pages/login';
+import Offer from '../../pages/offer';
+import NotFound from '../../pages/not-found';
+import Favorites from '../../pages/favorites';
+import PrivateRoute from '../private-route';
+import { useAppSelector } from '../../hooks';
+import LoadingScreen from '../../pages/loading-screen/loading-screen';
+import Layout from '../layout/layout';
+import { AuthStatus } from '../../utils/const';
 
 export default function App() {
   const authStatus = useAppSelector((state) => state.authStatus);
