+++ src/components/favorites-places/favorites-card.tsx
@@ -1,4 +1,4 @@
-import { OfferCardData } from '../types/types';
+import { OfferCardData } from '../../types/types';
 import { Link } from 'react-router-dom';
 
 type FavCardProps = {
