+++ src/components/favorites-places/favorites-places.tsx
@@ -1,4 +1,4 @@
-import { OfferCardData } from '../types/types';
+import { OfferCardData } from '../../types/types';
 import FavoritesCard from './favorites-card';
 
 
@@ -6,7 +6,7 @@ type FavPlacesProps = {
   offers: OfferCardData[];
 };
 
-export default function FavoritePlaces({offers}: FavPlacesProps) {
+export default function FavoritesPlaces({offers}: FavPlacesProps) {
   return(
     <div className="favorites__places">
       {
