+++ src/components/main-page/places-list/place-card.tsx
@@ -1,20 +1,23 @@
-import { OfferCardData } from '../types/types';
+import { MouseEvent } from 'react';
+import { OfferCardData } from '../../../types/types';
 import { Link } from 'react-router-dom';
-type NearCardProps = {
+
+type PlaceCardProps = {
   offer: OfferCardData;
+  onMouseEnter: (evt: MouseEvent<HTMLLIElement>) => void;
 }
 
-export default function NearPlacesCard({ offer }: NearCardProps) {
-  return (
-    <article className="near-places__card place-card">
+export default function PlaceCard({offer, onMouseEnter}: PlaceCardProps) {
+  return(
+    <article className="cities__card place-card" onMouseEnter={onMouseEnter} data-id={`${offer.id}`}>
       {offer.isPremium && (
         <div className="place-card__mark">
           <span>Premium</span>
         </div>
       )}
-      <div className="near-places__image-wrapper place-card__image-wrapper">
+      <div className="cities__image-wrapper place-card__image-wrapper">
         <Link to={`/offer/${offer.id}`}>
-          <img className="place-card__image" src={offer.previewImage} width="260" height="200" alt="Place image" />
+          <img className="place-card__image" src={offer.previewImage} width="260" height="200" alt="Place image"/>
         </Link>
       </div>
       <div className="place-card__info">
@@ -23,11 +26,11 @@ export default function NearPlacesCard({ offer }: NearCardProps) {
             <b className="place-card__price-value">&euro;{offer.price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button className="place-card__bookmark-button button" type="button">
+          <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
             <svg className="place-card__bookmark-icon" width="18" height="19">
               {offer.isFavorite && <use xlinkHref="#icon-bookmark"></use>}
             </svg>
-            <span className="visually-hidden">To bookmarks</span>
+            <span className="visually-hidden">In bookmarks</span>
           </button>
         </div>
         <div className="place-card__rating rating">
