+++ src/components/map/map.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from 'react';
 import { Icon, Marker, layerGroup} from 'leaflet';
 import 'leaflet/dist/leaflet.css';
-import useMap from '../hooks/use-map';
-import { Location, OfferCardData, OfferData} from '../types/types';
+import useMap from '../../hooks/use-map';
+import { Location, OfferCardData, OfferData} from '../../types/types';
 
 type MapProps = {
   mainPage: boolean;
