--- src/mocks/offers.ts
@@ -1,4 +1,0 @@
-export const Landlords = {
-  Max: {avatar: '/img/avatar-max.jpg', name: 'Max', isPro: true},
-  Angelina: {avatar: '/img/avatar-angelina.jpg', name: 'Angelina', isPro: false},
-};
